+++ Sources/Views/ALKChatCell.swift
@@ -29,6 +29,7 @@ public protocol ALKChatViewModelProtocol {
     var channelType: Int16 { get }
     var isMessageEmpty: Bool { get }
     var messageMetadata: NSMutableDictionary? { get }
+    var platformSource: String? { get }
 }
 
 extension ALKChatViewModelProtocol {
@@ -113,6 +114,24 @@ public final class ALKChatCell: SwipeTableViewCell, Localizable {
     public enum Config {
         public static var iconMuted = UIImage(named: "muted", in: Bundle.km, compatibleWith: nil)
     }
+    
+    let platformIcons: [String: UIImage] = [
+        "MOBILE": UIImage(named: "platformPhone", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "WEB": UIImage(named: "platformWeb", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "FACEBOOK": UIImage(named: "platformFacebook", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "WHATSAPPCLOUDAPI": UIImage(named: "platformWhatsapp", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "WHATSAPPTWILIO": UIImage(named: "platformWhatsapp", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "WHATSAPPDIALOG360": UIImage(named: "platformWhatsapp", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "DIALOG360": UIImage(named: "platform360Dialog", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "ZENDESKSUNSHINE": UIImage(named: "platformSunshine", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "VIBER": UIImage(named: "platformViber", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "LINE": UIImage(named: "platformLine", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "META": UIImage(named: "platformMeta", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "TELEGRAM": UIImage(named: "platformTelegram", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "MESSENGERFB": UIImage(named: "platformMessengerFB", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "INSTAGRAM": UIImage(named: "platformInstagram", in: Bundle.km, compatibleWith: nil) ?? UIImage(),
+        "TWILIO": UIImage(named: "platformTwillio", in: Bundle.km, compatibleWith: nil) ?? UIImage()
+    ]
 
     public var localizationFileName: String = "Localizable"
     var displayNames: ((Set<String>) -> ([String: String]?))?
@@ -136,7 +155,18 @@ public final class ALKChatCell: SwipeTableViewCell, Localizable {
         label.textColor = .text(.black00)
         return label
     }()
-
+    
+    private var platformImageView: UIImageView = {
+        let platformImage = UIImageView()
+        platformImage.contentMode = .scaleAspectFit
+        platformImage.clipsToBounds = true
+        
+        if !ALApplozicSettings.isAgentAppConfigurationEnabled(){
+            platformImage.isHidden = true
+        }
+        return platformImage
+    }()
+    
     private var messageLabel: UILabel = {
         let label = UILabel()
         label.lineBreakMode = .byTruncatingTail
@@ -266,7 +296,14 @@ public final class ALKChatCell: SwipeTableViewCell, Localizable {
     public func update(viewModel: ALKChatViewModelProtocol, identity _: ALKIdentityProtocol?, placeholder: UIImage? = nil) {
         self.viewModel = viewModel
         let placeHolder = placeholderImage(placeholder, viewModel: viewModel)
-
+        
+        /// this is used to set the icon according to data recived in metadata
+        if ALApplozicSettings.isAgentAppConfigurationEnabled(), let device = viewModel.platformSource, let platformImage = platformIcons[device] {
+            platformImageView.image = platformImage
+        } else {
+            platformImageView.image = nil
+        }
+        
         if let avatarImage = viewModel.avatarImage {
             if let imgStr = viewModel.avatarGroupImageUrl, let imgURL = URL(string: imgStr) {
                 let resource = ImageResource(downloadURL: imgURL, cacheKey: imgStr)
@@ -353,7 +390,7 @@ public final class ALKChatCell: SwipeTableViewCell, Localizable {
     }
 
     private func setupConstraints() {
-        contentView.addViewsForAutolayout(views: [avatarImageView, nameLabel, messageLabel, lineView, muteIcon, badgeNumberView, timeLabel, onlineStatusView, emailIcon])
+        contentView.addViewsForAutolayout(views: [avatarImageView, nameLabel, platformImageView, messageLabel, lineView, muteIcon, badgeNumberView, timeLabel, onlineStatusView, emailIcon])
         // setup constraint of imageProfile
         avatarImageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 17.0).isActive = true
         avatarImageView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 15.0).isActive = true
@@ -370,11 +407,21 @@ public final class ALKChatCell: SwipeTableViewCell, Localizable {
         emailIcon.heightAnchor.constraint(equalToConstant: Padding.Email.height).isActive = true
         emailIcon.leadingAnchor.constraint(equalTo: avatarImageView.trailingAnchor, constant: Padding.Email.left).isActive = true
         emailIcon.widthAnchor.constraintEqualToAnchor(constant: 0, identifier: ConstraintIdentifier.iconWidthIdentifier.rawValue).isActive = true
-
+        
         // setup constraint of mood'
         messageLabel.topAnchor.constraint(equalTo: nameLabel.bottomAnchor, constant: 2).isActive = true
         messageLabel.heightAnchor.constraint(equalToConstant: 20).isActive = true
-        messageLabel.leadingAnchor.constraint(equalTo: emailIcon.trailingAnchor, constant: 0).isActive = true
+        if platformImageView.isHidden {
+            messageLabel.leadingAnchor.constraint(equalTo: emailIcon.trailingAnchor, constant: 0).isActive = true
+        } else {
+            // setup constraint of Platform Identirier (Agent App)
+            platformImageView.topAnchor.constraint(equalTo: nameLabel.bottomAnchor, constant: 4).isActive = true
+            platformImageView.heightAnchor.constraint(equalToConstant: 15).isActive = true
+            platformImageView.leadingAnchor.constraint(equalTo: emailIcon.trailingAnchor).isActive = true
+            platformImageView.widthAnchor.constraint(equalToConstant: 15).isActive = true
+            
+            messageLabel.leadingAnchor.constraint(equalTo: platformImageView.trailingAnchor, constant: 6).isActive = true
+        }
         messageLabel.trailingAnchor.constraint(equalTo: muteIcon.leadingAnchor, constant: -8).isActive = true
 
         // setup constraint of line