+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/
 
 ## Unreleased
 - Added support for Sending GIF from device
+- Added icons for (mobile/web/facebook/WhatsApp) on conversation list for agent app. 
 
 ## [1.1.6] 2023-08-28
 - Fixed attachments upload issue 
@@ -13,9 +14,11 @@ The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/
 - Crash fix for agent app while updating it to latest ios SDK
 - Added validation and default value for form type rich message
 
+
 ## [1.1.4] 2023-08-11
 - Improved UI of multiple language selection & make it similar to android
 - Refresh Icon Change
+
 
 ## [1.1.3] 2023-08-08
 - Fixed down arrow coming in bottom of the screen when welcome message get rendered issue.